relocate directory components and rename DirectoryOrgs
