+++ src/components/club/directory/OrgDirectoryGrid.tsx
@@ -1,9 +1,9 @@
 import { type FC } from 'react';
-import DirectoryOrgs from './DirectoryOrgs';
+import DirectoryOrgs from '../ClubCard';
 import { api } from '@src/trpc/server';
 import { getServerAuthSession } from '@src/server/auth';
 import InfiniteScrollGrid from './InfiniteScrollGrid';
-import ScrollTop from './ScrollTop';
+import ScrollTop from '../../ScrollTop';
 
 interface Props {
   tag?: string;
